+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - The `main.php` script now resides directly in the `/src` root.
 - Errors are now reported in a friendlier form with more information about what may have gone wrong.
 
+## Fixed
+- A friendlier error message is now shown whenever the site files are deleted before the server is terminated.
+- Template content is now preprocessed through the foonoo parser to render foonoo tags.
+
 ## v0.5.0 - 2021-08-12
 ### Added
 - A `ContentOutputGenerated` event that fires whenever the output of a piece of content has been generated and is yet to be written to a file.